+++ src/test/java/org/swisspush/redisques/RedisQuesTest.java
@@ -49,6 +49,7 @@ public class RedisQuesTest extends AbstractTestCase {
 
     @After
     public void tearDown(TestContext context) {
+        QueueStatisticsCollector.CODECS_REGISTERED.set(false);
         vertx.close(context.asyncAssertSuccess());
     }